+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 	* Add ``WAVESubchunk``.
 * Keep unhandled WAVE subchunks.
 * Store ``ID3v2`` in ``FLAC`` in ``_id3`` attribute.
+* Emit warning when ignoring ID3v2 frames with bad values.
 
 ### Fixed